Fix WordPress Plugin Check warnings across the codebase
Sweep across admin, front, inc, templates, and uninstall to satisfy
WPCS / Plugin Check: translators comments, escaped exception/output,
unslash+sanitize on superglobals, gmdate/wp_rand/wp_parse_url/
wp_safe_redirect/wp_delete_file swaps, ABSPATH guards, LF line-ending
normalization, and documented phpcs:ignore reasons on intentional
direct-DB and discouraged-function calls. Correct TOC text-domain
to nuclear-engagement, add register_setting sanitize callbacks,
replace %i with esc_sql for WP 6.1 compatibility, and align README
plugin name with the header.

+++ nuclear-engagement/inc/Exceptions/ResourceException.php
@@ -57,7 +57,7 @@ class ResourceException extends BaseException {
 	 */
 	public static function executionTimeExceeded( float $time_percent, float $limit_percent ): self {
 		$max_execution = (int) ini_get( 'max_execution_time' );
-		$elapsed       = microtime( true ) - ( $_SERVER['REQUEST_TIME_FLOAT'] ?? microtime( true ) );
+		$elapsed       = microtime( true ) - (float) ( $_SERVER['REQUEST_TIME_FLOAT'] ?? microtime( true ) );
 
 		return new self(
 			sprintf( 'Execution time limit approaching: %.1f%% used (limit: %.0f%%)', $time_percent, $limit_percent ),
